Refactored query-network and query-network-status
