+++ modules/EventEmitter.js
@@ -309,21 +309,22 @@ class EventEmitter {
             });
         });
 
-        this._on('api-network-query', (data) => {
-            logger.info(`Network-query handling triggered with query ${JSON.stringify(data.query)}.`);
-
-            dvController.queryNetwork(data.query)
-                .then((queryId) => {
-                    data.response.status(201);
-                    data.response.send({
-                        message: 'Query sent successfully.',
-                        query_id: queryId,
-                    });
-                }).catch((error) => {
-                    logger.error(`Failed query network. ${error}.`);
-                    notifyError(error);
-                });
-        });
+        // Redundant code
+        // this._on('api-network-query', (data) => {
+        //     logger.info(`Network-query handling triggered with query ${JSON.stringify(data.query)}.`);
+        //
+        //     dvController.queryNetwork(data.query)
+        //         .then((queryId) => {
+        //             data.response.status(201);
+        //             data.response.send({
+        //                 message: 'Query sent successfully.',
+        //                 query_id: queryId,
+        //             });
+        //         }).catch((error) => {
+        //             logger.error(`Failed query network. ${error}.`);
+        //             notifyError(error);
+        //         });
+        // });
 
         this._on('api-choose-offer', async (data) => {
             const failFunction = (error) => {
@@ -367,37 +368,38 @@ class EventEmitter {
             }
         });
 
-        this._on('api-network-query-status', async (data) => {
-            const { id, response } = data;
-            logger.info(`Query of network status triggered with ID ${id}`);
-            const networkQuery = await Models.network_queries.find({ where: { id } });
-            if (networkQuery.status === 'FINISHED') {
-                try {
-                    const vertices = await dhService.dataLocationQuery(id);
-
-                    response.status(200);
-                    response.send({
-                        status: `${networkQuery.status}`,
-                        query_id: networkQuery.id,
-                        vertices,
-                    });
-                } catch (error) {
-                    logger.info(`Failed to process network query status for ID ${id}. ${error}.`);
-                    notifyError(error);
-                    response.status(500);
-                    response.send({
-                        error: 'Fail to process.',
-                        query_id: networkQuery.id,
-                    });
-                }
-            } else {
-                response.status(200);
-                response.send({
-                    status: `${networkQuery.status}`,
-                    query_id: networkQuery.id,
-                });
-            }
-        });
+        // Redundant code
+        // this._on('api-network-query-status', async (data) => {
+        //     const { id, response } = data;
+        //     logger.info(`Query of network status triggered with ID ${id}`);
+        //     const networkQuery = await Models.network_queries.find({ where: { id } });
+        //     if (networkQuery.status === 'FINISHED') {
+        //         try {
+        //             const vertices = await dhService.dataLocationQuery(id);
+        //
+        //             response.status(200);
+        //             response.send({
+        //                 status: `${networkQuery.status}`,
+        //                 query_id: networkQuery.id,
+        //                 vertices,
+        //             });
+        //         } catch (error) {
+        //             logger.info(`Failed to process network query status for ID ${id}. ${error}.`);
+        //             notifyError(error);
+        //             response.status(500);
+        //             response.send({
+        //                 error: 'Fail to process.',
+        //                 query_id: networkQuery.id,
+        //             });
+        //         }
+        //     } else {
+        //         response.status(200);
+        //         response.send({
+        //             status: `${networkQuery.status}`,
+        //             query_id: networkQuery.id,
+        //         });
+        //     }
+        // });
 
         this._on('api-offer-status', async (data) => {
             const { replicationId } = data;
